python/requirements.txt
Flask
pymongo
PyJWT